fix: moved DbUtilities temporary to repository/content. Might split it up
