+++ app/src/main/java/no/usn/mob3000/ui/screens/info/news/CreateNewsScreen.kt
@@ -14,7 +14,7 @@ import kotlinx.datetime.Clock
 import no.usn.mob3000.R
 import no.usn.mob3000.Viewport
 import no.usn.mob3000.data.model.content.NewsDto
-import no.usn.mob3000.data.network.DbUtilities
+import no.usn.mob3000.data.repository.content.DbUtilities
 import no.usn.mob3000.ui.theme.DefaultButton
 
 /**
