+++ app/src/main/java/no/usn/mob3000/ui/screens/info/news/UpdateNewsScreen.kt
@@ -11,12 +11,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
-import kotlinx.datetime.Clock
 import no.usn.mob3000.R
 import no.usn.mob3000.Viewport
 import no.usn.mob3000.data.model.content.NewsDto
-import no.usn.mob3000.data.network.DbUtilities
-import no.usn.mob3000.data.network.SupabaseClientWrapper
+import no.usn.mob3000.data.repository.content.DbUtilities
 import no.usn.mob3000.ui.theme.DefaultButton
 
 /**
